+++ tests/async/test-cancel-sync.R
@@ -21,9 +21,9 @@ test_that("detect, if one errors", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     async_detect(c(1 / 1000, 5), f)
   }
@@ -60,9 +60,9 @@ test_that("every, if one errors", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     async_every(c(1 / 1000, 5), f)
   }
@@ -82,9 +82,9 @@ test_that("filter, if one errors", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     async_filter(c(1 / 1000, 5), f)
   }
@@ -104,9 +104,9 @@ test_that("map, if one errors", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     async_map(c(1 / 1000, 5), f)
   }
@@ -143,9 +143,9 @@ test_that("every, if one errors", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     async_some(c(1 / 1000, 5), f)
   }
@@ -165,9 +165,9 @@ test_that("when_all, if one errors", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     defs <- lapply(c(1 / 1000, 5, 5), f)
     when_all(.list = defs)
@@ -208,9 +208,9 @@ test_that("when_some, if some error", {
   do <- function() {
     f <- function(n) {
       force(n)
-      delay(n)$then(function() pinged <<- pinged + 1)$then(
-        function() stop("foobar")
-      )
+      delay(n)$then(function() pinged <<- pinged + 1)$then(function() {
+        stop("foobar")
+      })
     }
     defs <- lapply(c(5, 1 / 1000, 1 / 1000, 5), f)
     when_some(3, .list = defs)