+++ tests/async/test-process.R
@@ -4,9 +4,9 @@ test_that("process", {
   afun <- async(function() {
     when_all(
       delay = delay(1 / 1000)$then(function(x) 1),
-      process = run_process(px, c("outln", "foobar"))$then(
-        function(x) str_trim(x$stdout)
-      ),
+      process = run_process(px, c("outln", "foobar"))$then(function(x) {
+        str_trim(x$stdout)
+      }),
       r_process = run_r_process(function() 2)$then(function(x) x$result)
     )
   })
@@ -26,9 +26,9 @@ test_that("process + http", {
     when_all(
       delay = delay(1 / 1000)$then(function() 1),
       http = http_get(http$url("/status/418"))$then(function(x) x$status_code),
-      process = run_process(px, c("outln", "foobar"))$then(
-        function(x) str_trim(x$stdout)
-      ),
+      process = run_process(px, c("outln", "foobar"))$then(function(x) {
+        str_trim(x$stdout)
+      }),
       r_process = run_r_process(function() 2)$then(function(x) x$result)
     )
   })